Forward UIKit logging notification to the WebContent process https://bugs.webkit.org/show_bug.cgi?id=277567 rdar://130763568
Reviewed by Ben Nham.

When blocking notifyd in the WebContent process sandbox, we need to forward notifications from the UI
process to the WebContent process instead. Otherwise, the WebContent process will not respond to this
notification. This patch forwards a UIKit logging notification, which is being observed in the
WebContent process.

* Source/WebKit/Resources/cocoa/NotificationAllowList/ForwardedNotifications.def:

Canonical link: https://commits.webkit.org/281827@main