Use OS_UNFAIR_LOCK_DATA_SYNCHRONIZATION and OS_UNFAIR_LOCK_ADAPTIVE_SPIN https://bugs.webkit.org/show_bug.cgi?id=246373 rdar://101055626
Reviewed by Saam Barati.

From the description of OS_UNFAIR_LOCK_DATA_SYNCHRONIZATION and OS_UNFAIR_LOCK_ADAPTIVE_SPIN[1], it looks good to use it in libpas.
So, this patch is just enabling them. Though we are not seeing any improvement / regression in benchmarks. Probably this is because
these locks are not contended anyway in libpas because libpas is designed to be extremely scalable (by deploying per-page locking when
it is contended).

[1]: https://opensource.apple.com/source/libplatform/libplatform-220.100.1/private/os/lock_private.h

* Source/bmalloc/libpas/src/libpas/pas_lock.h:
(pas_lock_lock):

Canonical link: https://commits.webkit.org/281849@main