Make `css-text-decor/invalidation/text-decoration-thickness.html` independent of viewport size https://bugs.webkit.org/show_bug.cgi?id=277653 rdar://133249123
Reviewed by Tim Nguyen.

Currently, this test can erroneously fail if the text is split into two lines and the following conditions are met;

1. The length of the second line is less than half the length of the first line
2. A click/hover happens in the center of the entire text

In this case, the click/hover does not actually end up over the text, and the test fails.

Fix by limiting the text's width so that it always wraps sufficiently, even on large viewport widths.

* LayoutTests/imported/w3c/web-platform-tests/css/css-text-decor/invalidation/text-decoration-thickness-expected.html:
* LayoutTests/imported/w3c/web-platform-tests/css/css-text-decor/invalidation/text-decoration-thickness-ref.html:
* LayoutTests/imported/w3c/web-platform-tests/css/css-text-decor/invalidation/text-decoration-thickness.html:

Canonical link: https://commits.webkit.org/281861@main