Remove the libavif project https://bugs.webkit.org/show_bug.cgi?id=277578 rdar://133097971
Reviewed by Tim Horton.

libavif was added to support AVIF images on macOS Monterey. The support for AVIF
was added to the system frameworks in macOS Ventura. Safari will not be released
anymore for macOS Monterey. ToT WebKit does not support macOS Monterey either.

Building libavif for WebKit is not needed anymore. Since dav1d is still needed
for webrtc, it will be moved outside libavif to be a child of PAL/ThirdParty.

* Source/WebCore/Configurations/WebCore.xcconfig:
* Source/WebCore/PAL/Configurations/PAL.xcconfig:
* Source/WebCore/PAL/PAL.xcodeproj/project.pbxproj:
* Source/WebCore/PAL/ThirdParty/dav1d: Renamed from Source/WebCore/PAL/ThirdParty/libavif/ThirdParty/dav1d.
* Source/WebCore/PAL/ThirdParty/libavif: Removed.
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/platform/network/cocoa/ResourceResponseCocoa.mm:
(WebCore::ResourceResponse::platformLazyInit):

Canonical link: https://commits.webkit.org/281902@main