Show the HTTPSOnly browsing warning https://bugs.webkit.org/show_bug.cgi?id=277554 rdar://133067048
Reviewed by Chris Dumez.

This change shows the browsing warning when a provisional navigation fails and
HTTPSOnly is active. Calling the delegates is now moved into a lambda so that
it can be easily shared with the warning's completion handler. This is helpful
for both testing and so the client can show more information, if appropriate,
when we call the delegate.

The new browsing warning contains a "Go Back" button and a "Continue" button,
but neither of them perform a specific action. The result of click either
button is calling the delegates, which is the existing behavior. I'll adjust
this behavior in a follow up patch.

I added and adjusted some API tests for this new warning.

* Source/WebKit/UIProcess/WebPageProxy.cpp:
(WebKit::WebPageProxy::didFailProvisionalLoadForFrameShared):
* Tools/TestWebKitAPI/Tests/WebKitCocoa/Navigation.mm:
(checkTitleAndClick):
(TEST(WKNavigation, HTTPSOnlyHTTPFallbackGoBack)):
(TEST(WKNavigation, HTTPSOnlyHTTPFallbackContinue)):
(TEST(WKNavigation, HTTPSOnlyWithSameSiteBypass)):
(TEST(WKNavigation, HTTPSOnlyWithHTTPRedirect)):
(TEST(WKNavigation, HTTPSOnlyHTTPFallback)): Deleted.

Canonical link: https://commits.webkit.org/281914@main