Add a paint behavior to force borders to black https://bugs.webkit.org/show_bug.cgi?id=277712 rdar://133328369
Reviewed by Darin Adler.

For `background-clip: border-area`, we'll need to sometimes paint a mask for the border clip.
To do this, we need to force the border colors to black when painting into the mask, so add a
paint behavior for this, much like the existing `ForceBlackText`.

The new behavior is not yet used.

* Source/WebCore/rendering/BorderEdge.cpp:
(WebCore::borderEdges):
* Source/WebCore/rendering/BorderEdge.h:
* Source/WebCore/rendering/BorderPainter.cpp:
(WebCore::BorderPainter::paintBorder):
* Source/WebCore/rendering/PaintPhase.h:
* Source/WebCore/rendering/RenderLayer.cpp:
(WebCore::RenderLayer::calculateClipRects const):

Canonical link: https://commits.webkit.org/281927@main