imported/w3c/web-platform-tests/IndexedDB/ready-state-destroyed-execution-context.html is failing https://bugs.webkit.org/show_bug.cgi?id=277753 rdar://133400072
Reviewed by NOBODY (OOPS!).

The test expects event to be dispatched to IDBRequest after its execution context is stopped (iframe is removed), and
this is not standard behavior. Browsers can choose to not dispatch events when script execution context stops, as in
Chrome and Safari nowadays. Update the test to wait on a new iframe with the same id to be loaded, instead of waiting on
event to be dispatched on IDBRequest. Also, since the goal of the test is to verify the accessor is valid after
execution context is destroyed (no exception is thrown), update the check to validate the type of the result, instead of
value of the result.

* LayoutTests/imported/w3c/web-platform-tests/IndexedDB/ready-state-destroyed-execution-context-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/IndexedDB/ready-state-destroyed-execution-context.html: