[JSC] Attempt to convert Move32 to Move in ARM64 https://bugs.webkit.org/show_bug.cgi?id=277744 rdar://133394615
Reviewed by Keith Miller.

We found that we are emitting Move32 unnecessarily when the destination said that the value is ZeroDefed.
We are doing so since x86_64 prefers Move32. But on ARM64, Move32 is more costly than Move. So we add opposite
optimization which attempts to make Move32 to Move (we already have optimization making Move to Move32 for x86_64).
We also fixes an issue in FTLOSRExitCompiler's logging introduced in 281831@main.

* Source/JavaScriptCore/b3/air/AirAllocateRegistersByGraphColoring.cpp:
* Source/JavaScriptCore/ftl/FTLOSRExitCompiler.cpp:
(JSC::FTL::compileStub):

Canonical link: https://commits.webkit.org/281959@main