REGRESSION (279847@main): Expanding threaded emails in Fastmail web UI results in missing content, repaint artifacts, broken rendering https://bugs.webkit.org/show_bug.cgi?id=277782 rdar://133344580
Reviewed by Tim Horton.

The optimization added in 279847@main caused us to skip repaints on a layer which moved, when the movement is triggered by
layout on an ancestor layer. This caused repaint bugs on Google Sheets when dragging rows, and Fastmail when collapsing
rows.

Fix by undoing the "simplified repaint" optimization on the descendant layers of a layer with overflow, by removing the check for
the layer also having a transform.

* LayoutTests/fast/repaint/simplified-repaint-with-overflow-expected.txt: Added.
* LayoutTests/fast/repaint/simplified-repaint-with-overflow.html: Added.
* Source/WebCore/rendering/RenderLayer.cpp:
(WebCore::RenderLayer::recursiveUpdateLayerPositions):

Canonical link: https://commits.webkit.org/281993@main