[UnifiedPDF] Two debug assertions on coordinate conversation fire unnecessarily for some locked documents https://bugs.webkit.org/show_bug.cgi?id=277814 rdar://133475258
Reviewed by NOBODY (OOPS!).

For some password protected documents, we get a zero page count when
locked. As such, the debug assertions touched by this patch fire
unnecessarily. We work around this by only doing the page index validity
check if we're not working with a document that exhibits this symptom.

* Source/WebKit/WebProcess/Plugins/PDF/UnifiedPDF/UnifiedPDFPlugin.h:
(WebKit::UnifiedPDFPlugin::convertDown const):
(WebKit::UnifiedPDFPlugin::convertUp const):