Check for runs of zero-bytes when linking AssemblerBuffer on x86 https://bugs.webkit.org/show_bug.cgi?id=276913 rdar://132273853
Reviewed by Yijia Huang.

The goal here is to detect memory corruption of the AssemblerBuffer
prior to link-time. On ARM64E, these would be caught at link-time via
the ARM64EHash feature within AssemblerBuffer, but due to that feature
being ARM64-specific, on x86 the link would succeed, leaving the
corruption to surface only when the code is executed, which makes it
much more difficult to identify the root cause.

* Source/JavaScriptCore/jit/ExecutableAllocator.h:
(JSC::performJITMemcpy):
* Source/WTF/wtf/PlatformEnable.h:

Canonical link: https://commits.webkit.org/282011@main