[JSC] Don't skip stress tests for `Intl.DurationFormat` on non-Darwin OS https://bugs.webkit.org/show_bug.cgi?id=277768
Reviewed by Yusuke Suzuki.

Currently, `stress/intl-durationformat-format-to-parts` is skipped on non-Darwin operating systems.
However, this test works on Linux, and other stress tests for `Intl.DurationFormat` are executed on
non-Darwin operating systems.

This patch changes the `stress/intl-durationformat-format-to-parts` test to run on non-Darwin
operating systems as well.

* JSTests/stress/intl-durationformat-format-to-parts.js:

Canonical link: https://commits.webkit.org/282014@main