[JSC] toUpperCase should check whether input string is already UpperCase https://bugs.webkit.org/show_bug.cgi?id=277845 rdar://problem/133513177
Reviewed by Alexey Shvayka.

Previously we thought toUpperCase against already-upper-cased-string is rare. So we had optimization for toLowerCase,
but didn't do that for toUpperCase. But in fact it is not rare, and there are many cases which does this. (e.g. tagName for example).

We should have a path not creating a new StringImpl for already UpperCase-ed string. This also allows JSC not to create a new JSString
around it since it can detect that new StringImpl is not created.

* Source/WTF/wtf/text/StringImpl.cpp:
(WTF::StringImpl::convertToUppercaseWithoutLocale):
(WTF::StringImpl::convertToUppercaseWithoutLocaleStartingAtFailingIndex8Bit):
(WTF::StringImpl::convertToUppercaseWithoutLocaleUpconvert):
* Source/WTF/wtf/text/StringImpl.h:

Canonical link: https://commits.webkit.org/282046@main