[@page] Enable support for margin descriptors https://bugs.webkit.org/show_bug.cgi?id=265314 rdar://118773100
Reviewed by Alan Baradlay.

Enable the feature flag; per
https://bugs.webkit.org/show_bug.cgi?id=200039, "Alan says this should
be enabled in shipping branches", but this seemingly never happened.

Rename the flag, since the only feature actually behind it is
support for the margin descriptor for @page; the rest of @page is
unconditionally enabled.

* LayoutTests/printing/page-with-10mm-left-margin.html:
* LayoutTests/printing/page-with-zero-margin.html:
* Source/WTF/Scripts/Preferences/UnifiedWebPreferences.yaml:
* Source/WebCore/page/PrintContext.cpp:
(WebCore::PrintContext::computedPageMargin):

Canonical link: https://commits.webkit.org/282048@main