Using Cross-Origin-Opener-Policy HTTP header disables the back/forward cache https://bugs.webkit.org/show_bug.cgi?id=274310 rdar://128678196
Reviewed by Per Arne Vollan.

When process-swapping on navigation response, due to the Cross-Origin-Opener-Policy HTTP header,
we would fail to create a SuspendedPageProxy for the page in the previous WebProcess, which
would essentially disable the back/forward cache.

This is not efficient and also causes issues when navigating back to a history item that was
created by JavaScript via history.pushState().

* Source/WebKit/UIProcess/WebPageProxy.cpp:
(WebKit::WebPageProxy::commitProvisionalPage):
* Tools/TestWebKitAPI/Tests/WebKitCocoa/ProcessSwapOnNavigation.mm:
(-[PSONNavigationDelegate _webView:navigation:didSameDocumentNavigation:]):
((ProcessSwap, NavigateBackAfterNavigatingAwayFromCOOPUsingBFCache)):

Canonical link: https://commits.webkit.org/282105@main