[Debug] `fast/multicol/crash-in-vertical-writing-mode.html` is a constant crash https://bugs.webkit.org/show_bug.cgi?id=277946
Reviewed by Alan Baradlay.

In some circumstances `RenderMultiColumnFlow::updateMinimumPageHeight()`
is being called when the fragments are still invalid and the asseration
`ASSERT(!m_fragmentsInvalidated)` is reached in
`RenderFragmentedFlow::fragmentAtBlockOffset()`.

To prevent this, we should check that
`RenderFragmentedFlow::hasValidFragmentInto()`.

This patch doesn't change the behaviour in the release build, but
prevents a crash in the debug build.

* LayoutTests/TestExpectations:
* Source/WebCore/rendering/RenderMultiColumnFlow.cpp:
(WebCore::RenderMultiColumnFlow::updateMinimumPageHeight):

Canonical link: https://commits.webkit.org/282127@main