Introducing CONJECTURE_ASSERT. https://bugs.webkit.org/show_bug.cgi?id=277938 rdar://133656316
Reviewed by Keith Miller.

CONJECTURE_ASSERT is only used to facilitate on-going analysis work to test conjectures about the code.
We want to be able to land these in the code base for some time to enable extended testing.

If the conjecture is proven false it, the CONJECTURE_ASSERT should either be removed or updated to test
a new conjecture.

If the conjecture is proven true, the CONJECTURE_ASSERT should either be promoted to an ASSERT or
RELEASE_ASSERT as appropriate, or removed if deemed of low value.

The number of CONJECTURE_ASSERTs should not be growing unboundedly, and they should not stay in the
code base perpetually.

* Source/WTF/wtf/Assertions.h:
* Source/WTF/wtf/PlatformEnable.h:

Canonical link: https://commits.webkit.org/282132@main